refactor: NoImageBox -> EmptySpaceBox로 변경
