+++ frontend/src/components/specific/emptySpaceBox/EmptySpaceBox.styles.ts
